Fix layout issues
